Added seeding

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\StorePost;
 use App\Models\Comment;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 //use Illuminate\Support\Facades\DB;
 
@@ -101,7 +102,7 @@ class PostsController extends Controller
         //abort_if(!isset($this->posts[$id]), 404);
 
         return view("posts.show", [
-            'post'=> BlogPost::with('comments')->findOrFail(5)
+            'post'=> BlogPost::with('comments')->findOrFail($id)
         ]);
     }
 
@@ -113,7 +114,11 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
-        return view('posts.edit',['post'=>BlogPost::findOrFail($id)]);
+        $post = BlogPost::findOrFail($id);
+
+        $this->authorize('update-post',$post);
+
+        return view('posts.edit',['post'=>$post]);
     }
 
     /**
@@ -126,6 +131,13 @@ class PostsController extends Controller
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+       /* if(Gate::denies('update-post',$post)){
+          abort(403, "You can't edit this post!");
+        }*/
+
+        $this->authorize('update-post',$post);
+
         $validated = $request->validated();
         $post->fill($validated);
         $post->save();
@@ -144,6 +156,9 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+        $this->authorize('update_post',$post);
+
         $post->delete();
         session()->flash('status','Blog post was deleted');
         return redirect()->route('posts.index');
